backend/requirements.txt
requests
beautifulsoup4
# mysql-connector-python
flask
flask-cors
python-dotenv
googlemaps